refactor(tests): split local corpus suites
