+++ AGENTS.md
@@ -37,10 +37,11 @@ The main development platform is macOS; keep default agent commands on this path
 - Run the self-contained CTest baseline:
   `ctest --test-dir QtProject/builds/build-debug-6.10.1-macos -C Debug --output-on-failure -R "^(test_comparison|test_mainwindow|test_repo_auto_delete|test_repo_video_matching)$"`
 - CTest labels make the safe lanes explicit: `ctest --test-dir QtProject/builds/build-debug-6.10.1-macos -L repo-fixtures` runs tracked fixtures; `-L local-fixtures` runs the optional `/Dev` matching suite; and `-L local-corpus` runs the `/Dev` corpus suite. `test_large_video_corpus` is separately labeled `large-corpus` and `requires-mounted-100gb`; run only its named functions.
-- Green `test_local_video_corpus` function cases for regular local-corpus work: `emptyDb`, `test_whole_app_nocache`, `test_whole_app_cached`, `test_whole_app_cache_only`.
+- Green `test_local_whole_app_scan` function cases for regular local-corpus work: `emptyDb`, `test_whole_app_nocache`, `test_whole_app_cached`, `test_whole_app_cache_only`.
 - `test_repo_video_extraction_regression` compares platform-sensitive metadata/thumbnails for the Nice videos; run it on the macOS dev setup, not Linux CI, unless refreshing reference expectations.
-- `test_local_video_corpus` whole-app and reference-detail cases depend on the `/Dev` fixture folder; run explicit functions only when that folder is available. Reference-detail cases (`test_check_refvidparams_nocache`, `test_check_refvidparams_withcache`, `test_check_refvidparams_withCacheOnly`) are not all green currently; run them when touching metadata, thumbnails, cache behavior, or reference data.
-- To investigate one `test_local_video_corpus` fixture, run a single data row with `test_function:data_tag`, for example `test_check_refvidparams_nocache:20150727_115225.mp4`.
+- `test_local_video_extraction_regression` compares every `/Dev` video against its metadata and thumbnail references in each cache mode. Reference-detail cases (`test_check_refvidparams_nocache`, `test_check_refvidparams_withcache`, `test_check_refvidparams_withCacheOnly`) are not all green currently; run them when touching metadata, thumbnails, cache behavior, or reference data.
+- `test_local_whole_app_scan` runs `/Dev` end-to-end scans in each cache mode; run its explicit functions only when that corpus is available.
+- To investigate one `/Dev` extraction reference, run a single `test_local_video_extraction_regression` data row, for example `test_check_refvidparams_nocache:20150727_115225.mp4`.
 - Prefer targeted `ctest --test-dir QtProject/builds/build-debug-6.10.1-macos ...` invocations when using CTest on macOS; the test CMake config chooses a platform plugin compatible with the current Qt build.
 - Do not run `test_large_video_corpus` unless explicitly requested; its active functions require the mounted 100GB folder.
 - Package macOS binaries: `npm run binaries`
