+++ QtProject/tests/test_local_whole_app_scan/tst_local_whole_app_scan.cpp
@@ -16,38 +16,13 @@ QDir developmentCorpus()
 }
 } // namespace
 
-class TestLocalVideoCorpus : public QObject
+class TestLocalWholeAppScan : public QObject
 {
     Q_OBJECT
 
   private slots:
     void initTestCase() { VideoCorpusTestHelpers::initialize(); }
     void emptyDb() { VideoCorpusTestHelpers::emptyDatabase(); }
-
-    void test_check_refvidparams_nocache_data()
-    {
-        VideoCorpusTestHelpers::addReferenceVideoRows(developmentCorpus(), 229);
-    }
-    void test_check_refvidparams_nocache()
-    {
-        VideoCorpusTestHelpers::verifyReferenceVideo(Prefs::NO_CACHE, true);
-    }
-    void test_check_refvidparams_withcache_data()
-    {
-        VideoCorpusTestHelpers::addReferenceVideoRows(developmentCorpus(), 229);
-    }
-    void test_check_refvidparams_withcache()
-    {
-        VideoCorpusTestHelpers::verifyReferenceVideo(Prefs::WITH_CACHE, true);
-    }
-    void test_check_refvidparams_withCacheOnly_data()
-    {
-        VideoCorpusTestHelpers::addReferenceVideoRows(developmentCorpus(), 229);
-    }
-    void test_check_refvidparams_withCacheOnly()
-    {
-        VideoCorpusTestHelpers::verifyReferenceVideo(Prefs::CACHE_ONLY, true);
-    }
 
     void test_whole_app_nocache()
     {
@@ -81,6 +56,6 @@ class TestLocalVideoCorpus : public QObject
     }
 };
 
-QTEST_MAIN(TestLocalVideoCorpus)
+QTEST_MAIN(TestLocalWholeAppScan)
 
-#include "tst_local_video_corpus.moc"
+#include "tst_local_whole_app_scan.moc"
